Added some Razor Views
Razor files are classes! Able to work with layouts and views. Layouts
are able to be used by all pages if called from the Layout section of
the view, or they could be done automatically to all if the _ViewStart
view is used to automatically fill out every page with the layout in
_Layout, so long as the individual View does not contain the Layout
section. HTML and Body are not required in this case.