Removed Angular, added some layouts and navigation
Trying to figure out what was wrong with the site, not showing on
intellisense properly. turns out they are dynamically typed, similar to
the viewbag of earlier. They are there, just have to know how to use it.
Might look into getting the resharper at a later point, but costs money.